+++ src/models/pyannote/mod.rs
@@ -1,4 +1,7 @@
-//! `PyAnnote` segmentation ONNX implementation.
+//! `PyAnnote` segmentation implementations and shared iterator.
 
+#[cfg(feature = "candle")]
+pub mod candle;
 pub mod iterator;
+#[cfg(feature = "onnxruntime")]
 pub mod onnx;